+++ .github/workflows/ci.yml
@@ -18,17 +18,4 @@ jobs:
     - run: npm -v
     - run: npm ci
     - run: npm run lint
-
-  test:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v3
-    - uses: volta-cli/action@v4
-    - run: npm ci
-    - run: npm link
-    - run: |
-        cd ./tests/fixtures/build-number
-        npm i
     - run: npm test
